refactor: extract Contata co-synthesis from synthesis entrypoint
Move the SMT obligation machinery and the mutual-group co-synthesis loop
out of the oversized aeon/synthesis/entrypoint.py (1097 lines) into a new
aeon/synthesis/modules/contata/cosynthesis.py, leaving entrypoint.py as a
thin orchestrator.

Moved: _value_literal, _z3_value, _z3_sort, _smt_unsat_core_obligations,
_trivial_stub, _joint_accepts, _accumulate_trace, _contains_member,
_find_top_level_def, _REL_BINOPS, _property_to_ir, _member_apps_in_ir,
_property_counterexamples, _refine_obligations, _contata_version_space_group,
_cosynthesize_group.

Kept in entrypoint.py (general orchestration): make_program, make_validator,
_make_fitness, make_evaluators, make_evaluator, make_output_evaluator,
_cluster_function, _synthesize_one, _mutual_group_ids, _partition_targets,
synthesize_holes, _pool_backed.

Dependency stays one-way: cosynthesis.py imports entrypoint._synthesize_one
via a function-local import to avoid a top-level cycle. entrypoint.py
re-exports _cosynthesize_group, _joint_accepts, _refine_obligations and
_smt_unsat_core_obligations (the last three are imported directly by the
mutual-synthesis tests).

Behavior-preserving: code moved verbatim, no signature changes.
